moved info_overlay into scrips updated imports
