+++ src/scripts/loot_filter_tts.py
@@ -3,7 +3,7 @@ import time
 from typing import TYPE_CHECKING
 
 import src.item.descr.read_descr_tts
-import src.info_overlay
+import src.scripts.info_overlay
 from src.cam import Cam
 from src.config.loader import IniConfigLoader
 from src.config.settings_models import ItemRefreshType, UnfilteredUniquesType
